raspberry-pi/requirements.txt
# Raspberry Pi detection dependencies
# Install: pip install -r requirements.txt

opencv-python-headless>=4.5.0   # Use opencv-python if not on Pi
numpy>=1.19.0
RPi.GPIO>=0.7.0                 # Only on Raspberry Pi (skip on other platforms)
